Clean up autoconf
I've done a lot of research on autoconf to try to make the
"automagically find remglk" better, and it looks like the consensus is
that one *shouldn't* try to make it "smart" that way.  There are a large
number of potential pitfalls with mismatched headers, etc.  A must more
reliable mechanism is "look only in the default locations, if it's not
there, require the user/installer to be explicit".

I've also added an actual linking check for remglk: we expect main() to
be provided by remglk, and it in turn requires certain functions to be
defined.